refactor: 데이터 전송만 담당하는 객체를 DTO 패키지로 이동
